属性应用逻辑部分

+++ crates/idns-core/src/utils/files.rs
@@ -3,7 +3,6 @@ use std::fs::File;
 use std::io::BufReader;
 use std::io::Read;
 use std::io::Write;
-use std::path::PathBuf;
 
 pub fn write_to_file(path: &str, filename: &str, content: &Vec<u8>) -> Result<()> {
     //
@@ -21,8 +20,10 @@ pub fn read_string_from_file(path: &str, filename: &str) -> Result<String> {
     //
     let storage_path = crate::utils::idns_home_path()?.join(path);
     std::fs::create_dir_all(storage_path.as_path())?;
-
     let filename = storage_path.join(filename);
+    if !filename.as_path().exists() {
+        return Err(anyhow!("文件不存在!"));
+    }
     let file = File::open(filename.as_path())?;
     let mut buf_reader = BufReader::new(file);
     // 解析配置文件
@@ -44,7 +45,9 @@ pub fn file_delete(path: &str, filename: &str) -> Result<bool> {
     let storage_path = crate::utils::idns_home_path()?.join(path);
     std::fs::create_dir_all(storage_path.as_path())?;
     let filename = storage_path.join(filename);
-
+    if !filename.as_path().exists() {
+        return Ok(true);
+    }
     std::fs::remove_file(filename)?;
     Ok(true)
 }
